PROD-2219 - learn code refactor
