+++ src-ts/tools/learn/learn-lib/lesson-provider/index.ts
@@ -2,4 +2,5 @@ export * from './learn-lesson-meta.model'
 export * from './learn-lesson.model'
 export * from './learn-module-meta.model'
 export * from './learn-module.model'
+export * from './lesson-provider-data.model'
 export * from './lesson.provider'
