+++ src-ts/tools/learn/learn-lib/lesson-provider/lesson-provider-data.model.ts
@@ -1,0 +1,7 @@
+import { LearnLessonMeta } from './learn-lesson-meta.model'
+
+export interface LessonProviderData {
+    lesson?: LearnLessonMeta
+    loading: boolean
+    ready: boolean
+}
