+++ src-ts/tools/learn/learn-lib/my-certifications-provider/index.ts
@@ -1,6 +1,7 @@
 export * from './learn-my-certification-progress.model'
 export * from './learn-my-certification.model'
 export * from './learn-my-module-progress.model'
+export * from './my-certification-progress-provider-data.model'
 export * from './my-certification-progress.provider'
 export * from './my-certifications-provider-data.model'
 export * from './my-certifications.provider'
