+++ src-ts/tools/learn/learn-lib/my-certifications-provider/my-certification-progress-provider-data.model.ts
@@ -1,0 +1,7 @@
+import { LearnMyCertificationProgress } from './learn-my-certification-progress.model'
+
+export interface MyCertificationProgressProviderData {
+    loading: boolean
+    progress?: LearnMyCertificationProgress
+    ready: boolean
+}
